moved test_conditional_independencies() to BaseEstimator class
